start with gcs auth
