+++ pr-deploy/package.json
@@ -16,11 +16,13 @@
     "test": "jest"
   },
   "devDependencies": {
+    "@types/jest": "^24.0.15",
+    "@types/node": "^12.0.8",
     "jest": "^24.8.0",
     "ts-jest": "^24.0.2",
     "typescript": "^3.5.2"
   },
   "dependencies": {
-    "@types/jest": "^24.0.15"
+    "google-auth-library": "^4.2.2"
   }
 }
